pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ig-follower-analyzer"
version = "0.1.0"
description = "Instagram follower analyzer and selective unfollow CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "playwright>=1.53.0,<2.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
]

[project.scripts]
ig-follower-analyzer = "ig_follower_analyzer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
